#13 - CR fixes
