+++ tests/Unit/EquipmentCastsTest.php
@@ -1,10 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Unit;
 
 use App\Models\Category;
 use App\Models\Equipment;
-use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
@@ -14,7 +15,6 @@ class EquipmentCastsTest extends TestCase
 
     public function testSavingEquipment(): void
     {
-        User::factory()->times(1)->create();
         Category::factory()->times(1)->create();
 
         $properties = [
@@ -38,12 +38,11 @@ class EquipmentCastsTest extends TestCase
                 "category_id" => Category::first()->id,
                 "serial_number" => "test",
                 "properties" => $properties,
-                "user_id" => User::first()->id,
             ]
         );
 
         $equipment = Equipment::first();
 
-        $this->assertEquals($properties, $equipment->properties->toArray());
+        $this->assertSame($properties, $equipment->properties->toArray());
     }
 }
